requirements_new.txt
opencv-python==4.11.0.86
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
Flask-SocketIO==5.3.6
easyocr==1.7.0
torch>=1.9.0
torchvision>=0.10.0
